Cart controller and entities relations

+++ src/main/java/com/ecommerce/abm_Daniel_Agresta/services/CartService.java
@@ -1,7 +1,9 @@
 package com.ecommerce.abm_Daniel_Agresta.services;
 
 import com.ecommerce.abm_Daniel_Agresta.entities.Cart;
+import com.ecommerce.abm_Daniel_Agresta.entities.Product;
 import com.ecommerce.abm_Daniel_Agresta.repositories.CartsRepository;
+import com.ecommerce.abm_Daniel_Agresta.repositories.ProductsRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -13,7 +15,17 @@ public class CartService {
     @Autowired
     private CartsRepository repository;
 
+    @Autowired
+    private ProductsRepository productRepository;
+
+    public Optional<Cart> cartById(Long id) {
+        return repository.findById(id);
+    }
+
     public Cart newCart(Cart cart) {
+        if (cart.getClient() == null) {
+            throw new IllegalArgumentException("Cart must been associated to a Client");
+        }
         return repository.save(cart);
     }
 
@@ -22,10 +34,56 @@ public class CartService {
     }
 
     public Optional<Cart> readCartById(Long id) {
+        if (id == null) {
+            throw new IllegalArgumentException("Cart id cannot be null");
+        }
         return repository.findById(id);
     }
 
+    public Cart addProductToCart(Long cartId, List<Product> products) {
+        Optional<Cart> cartOpt = repository.findById(cartId);
+        if (cartOpt.isPresent()) {
+            Cart cart = cartOpt.get();
+            for (Product product : products) {
+                product.getCarts().add(cart);
+                cart.getProducts().add(product);
+            }
+            cart.setAmount(cart.getProducts().size());
+            cart.setPrice(cart.getProducts().stream().mapToDouble(Product::getPrice).sum());
+            return repository.save(cart);
+        } else {
+            throw new IllegalArgumentException("Cart not founded");
+        }
+    }
+
+    public Cart addProductsToCart(Long cartId, List<Long> productIds) {
+        Optional<Cart> cartOpt = repository.findById(cartId);
+        if (cartOpt.isPresent()) {
+            Cart cart = cartOpt.get();
+            for (Long productId : productIds) {
+                Optional<Product> productOpt = productRepository.findById(productId);
+                if (productOpt.isPresent()) {
+                    Product product = productOpt.get();
+                    if (!cart.getProducts().contains(product)) {
+                        cart.getProducts().add(product);
+                        product.getCarts().add(cart);
+                    }
+                } else {
+                    throw new IllegalArgumentException("Product ID " + productId + ", not founded");
+                }
+            }
+            cart.setAmount(cart.getProducts().size());
+            cart.setPrice(cart.getProducts().stream().mapToDouble(Product::getPrice).sum());
+            return repository.save(cart);
+        } else {
+            throw new IllegalArgumentException("Cart not founded");
+        }
+    }
+
     public void deleteCart(Long id) {
+        if (id == null) {
+            throw new IllegalArgumentException("Cart id cannot be null");
+        }
         repository.deleteById(id);
     }
 }
